Rename hamilt_stodft to module_stodft
