+++ source/source_esolver/esolver_sdft_pw.h
@@ -2,10 +2,10 @@
 #define ESOLVER_SDFT_PW_H
 
 #include "esolver_ks_pw.h"
-#include "source_pw/hamilt_stodft/hamilt_sdft_pw.h"
-#include "source_pw/hamilt_stodft/sto_che.h"
-#include "source_pw/hamilt_stodft/sto_iter.h"
-#include "source_pw/hamilt_stodft/sto_wf.h"
+#include "source_pw/module_stodft/hamilt_sdft_pw.h"
+#include "source_pw/module_stodft/sto_che.h"
+#include "source_pw/module_stodft/sto_iter.h"
+#include "source_pw/module_stodft/sto_wf.h"
 
 namespace ModuleESolver
 {
