+++ source/source_pw/module_stodft/kernels/cuda/hpsi_norm_op.cu
@@ -1,4 +1,4 @@
-#include "source_pw/hamilt_stodft/kernels/hpsi_norm_op.h"
+#include "source_pw/module_stodft/kernels/hpsi_norm_op.h"
 
 #include <base/macros/macros.h>
 #include <cuda_runtime.h>
