+++ source/source_pw/module_stodft/sto_iter.h
@@ -3,7 +3,7 @@
 #include "source_base/math_chebyshev.h"
 #include "source_estate/elecstate_pw.h"
 #include "source_hamilt/hamilt.h"
-#include "source_pw/hamilt_stodft/hamilt_sdft_pw.h"
+#include "source_pw/module_stodft/hamilt_sdft_pw.h"
 #include "source_psi/psi.h"
 #include "sto_che.h"
 #include "sto_func.h"
